Add api dir, split action files
