+++ recipe-server/client/control/tests/actions/test_controlActions.js
@@ -2,8 +2,11 @@ import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import fetchMock from 'fetch-mock';
 
-import { fixtureRecipes, initialState } from '../fixtures.js';
-import * as actionTypes from '../../actions/ControlActions.js';
+import makeApiRequest from 'control/api';
+
+import { fixtureRecipes, initialState } from 'control/tests/fixtures';
+import * as actionTypes from 'control/actions/ControlActions';
+import { SHOW_NOTIFICATION } from 'control/actions/NotificationActions';
 
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
@@ -20,7 +23,7 @@ describe('controlApp Actions', () => {
     const expectedAction = { type: actionTypes.REQUEST_IN_PROGRESS };
     fetchMock.get('/api/v1/recipe/', fixtureRecipes);
 
-    return store.dispatch(actionTypes.makeApiRequest('fetchAllRecipes'))
+    return store.dispatch(makeApiRequest('fetchAllRecipes'))
     .then(() => {
       expect(store.getActions()).toContain(expectedAction);
     });
@@ -31,7 +34,7 @@ describe('controlApp Actions', () => {
       const expectedAction = { type: actionTypes.REQUEST_COMPLETE, status: 'success' };
       fetchMock.get('/api/v1/recipe/', fixtureRecipes);
 
-      return store.dispatch(actionTypes.makeApiRequest('fetchAllRecipes'))
+      return store.dispatch(makeApiRequest('fetchAllRecipes'))
       .then(() => {
         expect(store.getActions()).toContain(expectedAction);
       });
@@ -41,7 +44,7 @@ describe('controlApp Actions', () => {
       const expectedAction = { type: actionTypes.REQUEST_COMPLETE, status: 'error' };
       fetchMock.get('/api/v1/recipe/', { status: 500 });
 
-      return store.dispatch(actionTypes.makeApiRequest('fetchAllRecipes'))
+      return store.dispatch(makeApiRequest('fetchAllRecipes'))
       .catch(() => {
         expect(store.getActions()).toContain(expectedAction);
       });
@@ -49,7 +52,7 @@ describe('controlApp Actions', () => {
 
     it('creates a SHOW_NOTIFICATION action if provided', () => {
       const expectedAction = {
-        type: actionTypes.SHOW_NOTIFICATION,
+        type: SHOW_NOTIFICATION,
         notification: {
           messageType: 'error',
           message: 'Error fetching recipes.',
@@ -58,83 +61,9 @@ describe('controlApp Actions', () => {
       };
       fetchMock.get('/api/v1/recipe/', { status: 500 });
 
-      return store.dispatch(actionTypes.makeApiRequest('fetchAllRecipes')).catch(() => {
+      return store.dispatch(makeApiRequest('fetchAllRecipes')).catch(() => {
         expect(store.getActions()).toContain(expectedAction);
       });
-    });
-  });
-
-  it('makes a proper API request for fetchAllRecipes', () => {
-    fetchMock.get('/api/v1/recipe/', fixtureRecipes);
-
-    return store.dispatch(actionTypes.makeApiRequest('fetchAllRecipes'))
-    .then(() => {
-      expect(fetchMock.calls('/api/v1/recipe/').length).toEqual(1);
-    });
-  });
-
-  it('makes a proper API request for fetchSingleRecipe', () => {
-    fetchMock.get('/api/v1/recipe/1/', fixtureRecipes[0]);
-
-    return store.dispatch(actionTypes.makeApiRequest('fetchSingleRecipe', { recipeId: 1 }))
-    .then(() => {
-      expect(fetchMock.calls('/api/v1/recipe/1/').length).toEqual(1);
-    });
-  });
-
-  it('makes a proper API request for fetchSingleRevision', () => {
-    fetchMock.get('/api/v1/recipe_version/169/', fixtureRecipes[0]);
-
-    return store.dispatch(actionTypes.makeApiRequest('fetchSingleRevision', { revisionId: 169 }))
-    .then(() => {
-      expect(fetchMock.calls('/api/v1/recipe_version/169/').length).toEqual(1);
-    });
-  });
-
-  it('makes a proper API request for addRecipe', () => {
-    fetchMock.post('/api/v1/recipe/', fixtureRecipes[0]);
-
-    return store.dispatch(actionTypes.makeApiRequest('addRecipe', { recipe: fixtureRecipes[0] }))
-    .then(() => {
-      const calls = fetchMock.calls('/api/v1/recipe/');
-      expect(calls[0][1].body).toEqual(JSON.stringify(fixtureRecipes[0]));
-    });
-  });
-
-  it('makes a proper API request for updateRecipe', () => {
-    fetchMock.mock('/api/v1/recipe/1/', fixtureRecipes[0], { method: 'PATCH' });
-
-    return store.dispatch(actionTypes.makeApiRequest('updateRecipe', {
-      recipe: fixtureRecipes[0], recipeId: 1,
-    }))
-    .then(() => {
-      const calls = fetchMock.calls('/api/v1/recipe/1/');
-      expect(calls[0][1].body).toEqual(JSON.stringify(fixtureRecipes[0]));
-    });
-  });
-
-  it('makes a proper API request for deleteRecipe', () => {
-    fetchMock.delete('/api/v1/recipe/1/', fixtureRecipes[0]);
-
-    return store.dispatch(actionTypes.makeApiRequest('deleteRecipe', { recipeId: 1 }))
-    .then(() => {
-      expect(fetchMock.calls('/api/v1/recipe/1/').length).toEqual(1);
-    });
-  });
-
-  describe('showNotification', () => {
-    it('automatically dismisses notifications after 10 seconds', async () => {
-      jasmine.clock().install();
-
-      const notification = { messageType: 'success', message: 'message' };
-      await store.dispatch(actionTypes.showNotification(notification));
-
-      const dismissAction = actionTypes.dismissNotification(notification.id);
-      expect(store.getActions()).not.toContain(dismissAction);
-      jasmine.clock().tick(10001);
-      expect(store.getActions()).toContain(dismissAction);
-
-      jasmine.clock().uninstall();
     });
   });
 });
